backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.0
gunicorn==22.0.0
